requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0